SizePrecalculation logic changed, added social media platform icons and added logic for the ability to upload videos with certain parameters to the platform

+++ VideoConverterApi/Models/PrecalculatedSize.cs
@@ -2,6 +2,6 @@
 
 public class PrecalculatedSize
 {
-    public int SizeInBits { get; set; }
+    public ulong SizeInBits { get; set; }
     public double SizeInMegaBytes { get; set; }
 }
